requirements.txt
torch
huggingface_datasets
tokenizers
tqdm
numpy